Provide backwards compatible Hoptoad support
