+++ Gemfile
@@ -9,4 +9,9 @@ group :test do
   gem "leftright", :platforms => :mri_18
   gem "yajl-ruby", "~>0.8.2", :platforms => :mri
   gem "json", "~>1.5.3", :platforms => [:jruby, :rbx]
+
+  # Needed to run airbrake/hoptoad tests
+  # gem "i18n"
+  # gem "airbrake"
+  # gem "hoptoad_notifier"
 end
